% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{api_url}
\alias{api_url}
\title{Get or set the api url for all operations.}
\usage{
api_url(url)
}
\arguments{
\item{url}{the api url to be used. The api url is returned invisibly.}
}
\value{
Either the old api url (when \code{url} is supplied) or the current api url (when \code{url} is not supplied)
}
\description{
Get or set the api url for all operations.
}
\examples{
api_url()

}
