% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_helper.R
\name{as.hash}
\alias{as.hash}
\title{Convert x into hash}
\usage{
as.hash(x)
}
\arguments{
\item{x}{a vector which can be, by using \code{as.character(x)}, converted to a character vector of length 1}
}
\value{
the to a character converted object x with the class \code{hash} assigned
}
\description{
When \code{length(x) == 1} and x can be converted to a character by using the
\code{as.character()} function, x is converted to a hash object. This does
\bold{not} imply an=y hashing, just type conversion!
}
\details{
Only really useful to convert a hash represented as a string into an object of class \code{hash}.
}
\examples{
as.hash("2c5d36be542f8f0e7345d77753a5d7ea61a443ba6a9a86bb060332ad56dba38e")

}
