% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{content_is_OK}
\alias{content_is_OK}
\title{Check and raise error}
\usage{
content_is_OK(OriginStampResponse)
}
\arguments{
\item{OriginStampResponse}{object of type \code{OriginStampResponse}}
}
\value{
error message or \code{TRUE} if no error.
}
\description{
Check and raise error
}
