% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timestamp.R
\name{create_timestamp}
\alias{create_timestamp}
\title{Create a timestamp for a hash}
\usage{
create_timestamp(
  hash,
  error_on_fail = TRUE,
  comment = "None",
  notifications = list(currency = 0, notification_type = 0, target =
    "originstamp@trashmail.com")
)
}
\arguments{
\item{hash}{hash for which shuld be sum=bmitted to OroginStamp}

\item{error_on_fail}{if \code{TRUE}, raise error when api call fails, otherwise return the failed response.}

\item{comment}{a comment for the new timestamp}

\item{notifications}{notification settings}
}
\value{
object of type \code{OriginStampResponse}
}
\description{
Submit a hash to obtain a Trusted Time Stamp for that hash
wrapper around \url{https://doc.originstamp.org/#!/default/post_hash_string}
}
\examples{
  \dontrun{
    # Store hash
    store_hash(
      hash = "2c5d36be542f8f0e7345d77753a5d7ea61a443ba6a9a86bb060332ad56dba38e",
      information = list(name = "My Name", comments = "A fantastic example")
    )
  }
}
