% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{exists_option}
\alias{exists_option}
\title{Check if option exists}
\usage{
exists_option(name)
}
\arguments{
\item{name}{name of option}
}
\value{
\code{logical} indicating if option \code{name} exists
}
\description{
Check if option exists
}
