% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_content.R
\name{extract_content}
\alias{extract_content}
\title{Extract and convert content from response}
\usage{
extract_content(response)
}
\arguments{
\item{response}{response as returned from \code{store_hash()} and \code{get_hash_info()}}
}
\value{
content in JSON
}
\description{
Convert timestamps to human readable date
}
