% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hash.R
\name{extract_hash}
\alias{extract_hash}
\title{Return the hash of an xml proof or pdf certificate}
\usage{
extract_hash(x)
}
\arguments{
\item{x}{filename or URL pointing either to the xml proof or the pdf certificate issued from OriginStamp.}
}
\value{
an \code{hash} object containing the hash as used in the proof
}
\description{
Return the hash of an xml proof or pdf certificate
}
\examples{
fn <- "https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/proof.Bitcoin.xml"
extract_hash(fn)
fn <- "https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf"
extract_hash(fn)

}
