% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_root_hash.R
\name{extract_root_hash}
\alias{extract_root_hash}
\title{Return the root hash of an xml proof or pdf certificate}
\usage{
extract_root_hash(x)
}
\arguments{
\item{x}{filename or URL pointing either to the xml proof or the pdf certificate issued from OriginStamp.}
}
\value{
an \code{hash} object containing the root hash
}
\description{
Return the root hash of an xml proof or pdf certificate
}
\examples{
hn <- "https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/proof.Bitcoin.xml"
extract_root_hash(hn)

hn <- "https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf"
extract_root_hash(hn)

}
