% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hash.R
\name{extract_transaction}
\alias{extract_transaction}
\title{Return the transaction id from the pdf certificate}
\usage{
extract_transaction(x)
}
\arguments{
\item{x}{filename or URL pointing either to the the pdf certificate issued from OriginStamp.}
}
\value{
the transaction id
}
\description{
Return the transaction id from the pdf certificate
}
\examples{
transaction("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf")

}
