% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractContent.R
\name{extractContent}
\alias{extractContent}
\title{Extract and convert content from response}
\usage{
extractContent(response)
}
\arguments{
\item{content}{response as returned from \code{store_hash()} and \code{get_hash_info()}}
}
\value{
content in JSON
}
\description{
Convert timestamps to human readable date
}
