% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complete_seed_file.R
\name{get_complete_seed_file}
\alias{get_complete_seed_file}
\title{Get complete Seed File}
\usage{
get_complete_seed_file(hash, file = "", error_on_fail = TRUE)
}
\arguments{
\item{hash}{hast from which to download the seed}

\item{file}{file name to store the seed to. If \code{NULL}, save to default name (\code{HASH.OriginStamp.seed.txt}), if \code{""} (default) do not save, otherwise to \code{file}.}

\item{error_on_fail}{if TRUE, raise error when api call fails, otherwise return the failed response.}
}
\value{
object of class \code{OriginStampResponse}, \code{content} contains seed
}
\description{
wrapper around https://doc.originstamp.org/#!/default/get_download_seed_hash_string
}
