% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currencies.R
\name{get_currencies}
\alias{get_currencies}
\title{Get the active currencies}
\usage{
get_currencies(error_on_fail = TRUE)
}
\arguments{
\item{error_on_fail}{if TRUE, raise error when api call fails, otherwise
return the failed response.}
}
\value{
object of class \code{OriginStampResponse}.
}
\description{
wrapper around \url{https://api.originstamp.com/swagger/swagger-ui.html#/scheduler/getActiveCurrenciese}. The
function downloads information about the currencies.
}
\examples{
  \dontrun{
    # Retrieve complete merkle tree proof
    get_currencies()
  }
}
