% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_helper.R
\name{hash}
\alias{hash}
\alias{hash.file}
\alias{hash.character}
\title{Calculate the hash from a character vector or vector of file names}
\usage{
hash(x)

hash.file(x)

hash.character(x)
}
\description{
If all elements point to existing files, the hashes of these files
  are calculated, otherwise the hashes of the character vector, one per element.

Calculates the sha256 hash of each file in \code{x}.

Calculates the sha256 hash of each element in \code{x}.
}
