% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_helper.R
\name{hash_file}
\alias{hash_file}
\title{Calculate sha256 hash from file}
\usage{
hash_file(file)
}
\arguments{
\item{file}{filename}
}
\value{
sha256 hash of the file
}
\description{
A helper function to calculate the sha256 hash from a file.
This can be a binary or text file.
}
\examples{
hash_file( system.file( "DESCRIPTION", package = "ROriginStamp") )

}
