% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_OriginStampResponse.R
\name{new_OriginStampResponse}
\alias{new_OriginStampResponse}
\title{New \code{OriginStampResponse} object}
\usage{
new_OriginStampResponse()
}
\value{
object of class \code{OriginStampResponse}
}
\description{
Create new \code{OriginStampResponse} object. The object contains the following fields:
\describe{
  \item{content}{The returned object, converted into a format suitable for R. The exact format depends on the type of request.}
  \item{response}{The actual respons from the call via the \code{httr} package as an object of type \code{response}.}
}
}
\examples{
x <- new_OriginStampResponse()
x
# $content
# [1] NA
#
# $response
# [1] NA
#
# attr(,"class")
# [1] "OriginStampResponse"
}
