% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_hash.R
\name{root_hash}
\alias{root_hash}
\title{Return the root hash of an xml proof or pdf certificate}
\usage{
root_hash(x, verify = FALSE)
}
\arguments{
\item{x}{filename or URL pointing either to the xml proof or the pdf certificate issued from OriginStamp.}

\item{verify}{if \code{FALSE}, read the root hash from the xml file. If \code{TRUE}, verify the root hash using the proof.}
}
\value{
the root hash
}
\description{
Return the root hash of an xml proof or pdf certificate
}
\examples{
root_hash("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/proof.Bitcoin.xml")
root_hash("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf")

}
