% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROriginStamp-package.R
\docType{package}
\name{ROriginStamp-package}
\alias{ROriginStamp}
\alias{ROriginStamp-package}
\title{ROriginStamp: Interface to OriginStamp API to Obtain Trusted Time Stamps}
\description{
Functions to get and download trusted timestamps and seed files from OriginStamp <https://originstamp.com>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rkrug/ROriginStamp}
  \item Report bugs at \url{https://github.com/rkrug/ROriginStamp/issues}
}

}
\author{
\strong{Maintainer}: Rainer M. Krug \email{Rainer.Krug@uzh.ch} (\href{https://orcid.org/0000-0002-7490-0066}{ORCID})

}
\keyword{internal}
