% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROriginStamp_options.R
\name{ROriginStamp_options}
\alias{ROriginStamp_options}
\title{Set or get options for my ROriginStamp package}
\usage{
ROriginStamp_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my ROriginStamp package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{api_url}} {(\code{character} = \code{"https://api.originstamp.org/api/"}) The url of the OriginStamp API }
 \item{\code{api_key}} {(\code{numeric} = \code{"Please get valid API key from Originstamp.org!"}) The api key.  This needs to be obtained from \link{https://originstamp.org/} }
}
}

