% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_option}
\alias{set_option}
\title{Set Option}
\usage{
set_option(name, value)
}
\arguments{
\item{name}{name of option}

\item{value}{value for option}
}
\value{
the old value from that Option
}
\description{
Set Option
}
\examples{

}
