% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_hash.R
\name{store_hash}
\alias{store_hash}
\title{Storing a hash}
\usage{
store_hash(hash, error_on_fail = TRUE, information = list(email = "none",
  name = "NULL", comment = "Just a test"))
}
\arguments{
\item{hash}{hash for which shuld be sum=bmitted to OroginStamp}

\item{error_on_fail}{if \code{TRUE}, raise error when api call fails, otherwise return the failed response.}

\item{information}{contains information which is stored together with the submitted hash_string. Has to be a named list with each element being of length one}
}
\value{
object of type \code{OriginStampResponse}
}
\description{
Store a hash to obtain a Trusted Time Stamp
wrapper around \url{https://doc.originstamp.org/#!/default/post_hash_string}
}
