% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txid_from _root_hash.R
\name{txid_from_root_hash}
\alias{txid_from_root_hash}
\title{Title}
\usage{
txid_from_root_hash(root_hash)
}
\arguments{
\item{root_hash}{root hash which has been submitted as OP_RETURN}
}
\value{
the transaction id
}
\description{
Title
}
