% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_verify_proof.R
\name{verify_proof}
\alias{verify_proof}
\title{Verify the proof}
\usage{
verify_proof(proof, verbose = FALSE)
}
\arguments{
\item{proof}{}
}
\value{
\code{TRUE} if verified successfully, otherwise \code{FALSE}
}
\description{
INTERNAL FUNCTION!
Verify the proof. The proof needs to be provided as an xml object
}
