% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{api_key}
\alias{api_key}
\title{Get or set the api key for all operations.}
\usage{
api_key(key)
}
\arguments{
\item{key}{the api key to be used. The old api key is returned invisibly.}
}
\value{
Either the old api key (when \code{key} is supplied) or the current api key (when \code{key} is not supplied)
}
\description{
Get or set the api key for all operations.
}
\examples{
api_key()

}
