% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_proof.R
\name{extract_proof}
\alias{extract_proof}
\title{Extract the proof from an xml proof or pdf certificate}
\usage{
extract_proof(x, verify = FALSE)
}
\arguments{
\item{x}{filename or URL pointing either to the xml proof or the pdf certificate issued from OriginStamp.}

\item{verify}{if \code{FALSE}, read the root hash from the xml file. If \code{TRUE}, verify the root hash using the proof.}
}
\value{
the root hash
}
\description{
If the v arification fails, the function aborts and raises an error.
}
\examples{
extract_proof("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/proof.Bitcoin.xml")
extract_proof("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf")

}
