% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hash_status.R
\name{get_hash_status}
\alias{get_hash_status}
\title{Retrieve hash Information}
\usage{
get_hash_status(hash, error_on_fail = TRUE)
}
\arguments{
\item{hash}{hast from which to download the info}

\item{error_on_fail}{if \code{TRUE}, raise error when api call fails, otherwise return the failed response.}
}
\value{
object of type \code{OriginStampResponse}, \code{content} contains the additional info as \code{list}.
}
\description{
wrapper around \url{https://doc.originstamp.org/#!/default/getHashInformation}
}
\examples{
  \dontrun{
    # get hash info
    get_hash_status(
      hash = "2c5d36be542f8f0e7345d77753a5d7ea61a443ba6a9a86bb060332ad56dba38e"
    )
  }
}
