% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_OriginStampResponse.R
\name{new_OriginStampResponse}
\alias{new_OriginStampResponse}
\title{New \code{OriginStampResponse} object}
\usage{
new_OriginStampResponse(response = NULL)
}
\arguments{
\item{response}{a response from a call to \code{curl::curl_fetch_memory()} which is used to fill the return values of the new object}
}
\value{
object of class \code{OriginStampResponse}
}
\description{
Create new \code{OriginStampResponse} object. The object contains the following fields:
\describe{
  \item{status}{TThe status code of the request. 200 if OK. see \url{https://api.originstamp.com/swagger/swagger-ui.html} for details.}
  \item{content}{The returned object, converted into a format suitable for R. The exact format depends on the type of request.}
  \item{headers}{The returned eaders.}
}
}
\examples{
x <- new_OriginStampResponse()
x
# $content
# [1] NA
#
# $response
# [1] NA
#
# attr(,"class")
# [1] "OriginStampResponse"
}
