% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_hash.R
\name{store_hash}
\alias{store_hash}
\title{Storing a hash}
\usage{
store_hash(hash, error_on_fail = TRUE, information = NULL)
}
\value{
object of type \code{OriginStampResponse}
}
\description{
Store a hash to obtain a Trusted Time Stamp
wrapper around https://doc.originstamp.org/#!/default/post_hash_string
}
